Add more introspection rules - local boot, LLDP description to name
We currently hard code the interface from which to pull the LLDP port description.
This should be improved at some point, but we are limited by indexing the interfaces
dict by some constant. We could provide a rule for each possible interface name
but this feels ugly. Alternatively we could do this with a custom plugin.