Support existing flavors in compute node flavor registration
Previously, an unregistered flavor which sorts to a position before an existing
registered flavor (due to its attributes being 'smaller') would take the name
of the existing flavor and register a duplicate flavor for the registered one.

This change takes account of the existing registered flavors when determining
the flavors to register.

This still has the limitation that it does not work for existing flavors that
are not consecutively numbered from 0.