Fix issue with adding user to docker group
In order to activate the group membership for the user it is necessary
to log out and back in again. Ansible typically uses the SSH Control
Persist feature which will keep the user logged in between successive
tasks. This change stops the control persist connection when the user
has been added to the docker group.

The method for determining the path to the control socket is a little
more hard coded than it should be, but no other method could be found
for determining the socket's path.