Fix defaults for environment variable lookups
The env lookup plugin returns None if the environment variable is not
defined, which does not count as undefined. This means the default filter
will not replace the value. We pass true as the second argument to default
in order to interpret the value as a boolean when determining if it is
defined.