Add the ability to call --list-tasks
This allows you to print out which tasks will be executed, without
actually running the tasks. It could help debug using the tag and
filtering features. When developing, it makes the feedback loop quicker
because non of the tasks need to execute.

Note this doesn't work for kolla-ansible, as their CLI doesn't support
this command line. Which frankly makes this a little useless, but we
could always push that upstream if it proves handy.