Fix image build regex
Turns out kolla_docker_registry is always defined, but is sometimes an
empty string or None. It defaults to None.

Before this change we always sent "--registry" but the kolla build CLI
would pick up our regex as the parameter associated with "--registry" so
we would get a strange image tag, and still build all the images.