Use largest MTU of interfaces when determining OVS veth MTU
We create a veth pair to link the Linux bridge to the neutron Open vSwitch
bridge.  The MTU of the veth needs to be the largest MTU of the networks that
it will carry.  When VLAN subinterfaces are used, the provisioning network may
map to the same bridge as one of the external networks, and we must use the
largest MTU of these networks.