Install pip when not using a kayobe target virtualenv
Previously we were relying on kolla-ansible's baremetal role to install pip, but if
virtualenvs are used for the target hosts with kolla-ansible (but not kayobe), then
pip will not be installed to the system site packages. This change installs pip
using easy_install if target virtualenvs are not in use for kayobe.