Fix undefined variable in cadvisor role
When running latest kayobe on alaska, needed to ensure the correct
monitoring configuration was applied, I hit undefined variables,
hitting errors such as:

  The task includes an option with an undefined variable.
  The error was: 'nodeexporter_restart_policy' is undefined

On inspecting the role, it seems some cut and paste errors from the
nodeexporter role have snook in. The command is not required for the
docker_container directive, its not specified in the defaults.

Interestingly this triggered when I was just trying to redeploy haproxy.