Add initial in-repo Zuul jobs
This change ensures parity with the CI jobs run by TravisCI on the
legacy StackHPC github repo.

In addition to the system-required job in infra project-config, the
following jobs are run on check and gate:

- openstack-tox-pep8
- openstack-tox-py27
- build-openstack-sphinx-docs
- kayobe-tox-ansible-syntax
- kayobe-tox-ansible

Ansible molecule tests are not working currently, as they require docker
engine to be running.

Change-Id: Ie95bcdaabb10ab661bc3a01efbaf9143896d6913
Story: #2001637
Task: #6646