Make kayobe ansible user bootstrap optional
The bootstrap user may be used to create the kayobe user account and
configure passwordless sudo. We can't assume that the bootstrap user
account will exist after the initial bootstrapping, or that the current
operator's key is authorised for the bootstrap user. We therefore
attempt to access the kayobe user account via SSH, and only perform the
bootstrap process if the account is inaccessible.

This change also adds some tasks to verify that the kayobe ansible user
is accessible and has passwordless sudo configured.

Change-Id: Ibdab0053caa2db71df2fd03cc8a598ae5aac73c9
Story: 2001659
Task: 6692