Configure SELinux state in kolla ansible
Kolla ansible now provides the ability to change the SELinux mode, and does so
by default. The default mode is 'permissive', whereas kayobe sets it to
'disabled' in the disable-selinux role.  This results in a flip-flop effect as
the two fight, and worse - kayobe will reboot the system to apply the change
on subsequent runs of 'kayobe <seed|controller> host configure'.

This change configures the selinux mode for kolla ansible to be 'disabled' to
avoid this issue.

TrivialFix

Change-Id: I53e1d431ecd5ddb602f41b197ac482c3ed89d1d9