Fix handling of switch_type variable in Ansible 2.5
In Ansible, playbook group variables take precedence over inventory
group variables. Typically in a kayobe-config we might specify the
switch_type variable for a group containing switches using inventory
group_vars. This seemed to work in Ansible 2.3, but since Ansible 2.4
the default value of None in kayobe's playbook group_vars takes
precedence. This behaviour is as documented, not sure why it worked
previously.

This change removes the default value of None set in the playbook
group_vars.

Change-Id: I042ad8d115cdc0dd57a116e1be9d397ef4d216ae
Story: 2002903
Task: 22870