Fixes for baremetal compute commands
Gather facts prior to using the stackhpc.os-shade role. This is required
following a recent change in that role to reference the ansible_user_uid and
ansible_user_gid facts.

Allow 'kayobe baremetal compute rename' command to ignore compute nodes in the
ansible inventory but not found in ironic.

Change-Id: I75a2f0dd0192e41de7b91376bb32b4149df04a60
TrivialFix