Merge pre-reno release notes with reno notes
Prior to the Queens release, release notes were added to the kayobe
sphinx documentation. During the Queens release cycle they were changed
to be generated via the reno tool.

This change moves the pre-Queens release notes to the reno source,
ensuring that all release notes are available in one place.

Change-Id: I9606cbe50aacc498a645bff163e3d2dafac198e3