Fix use of --ask-vault-pass argument
Currently, this argument does not work correctly, since the vault
password is not passed to kayobe via $KAYOBE_VAULT_PASSWORD, meaning
that it cannot update the kolla-ansible passwords.yml file.

It also works non-optimally, since every invocation of ansible-playbook
will prompt for a password.

This change fixes the issue by prompting for a password once in the
kayobe CLI, and storing the password in the $KAYOBE_VAULT_PASSWORD
environment variable. The kayobe-vault-password-helper command is then
used as the --vault-password-file to ansible-playbook, and the helper
pulls the password out of the environment.

Change-Id: I88b1b7b9e9be15b52e730d353ce1b1a6feacceb8
Story: 2001664
Task: 27009
(cherry picked from commit 358b5c68828e6e7de8bfc7654c7418c02380052b)