Summary should be a one liner
Without this change twine check dist/* would fail as this was
producing a double new line. The rest of the PKG-INFO would then
be interpreted as the message body and become the long
description. See https://github.com/pypa/twine/issues/422 for
more details.

This fixes the test-release-openstack-python3 zuul job.

TrivialFix

Change-Id: If1330c4824e949aa427475dfcc26218fb398a66e