Automatically enable serial consoles on post configure
This functionality is controlled by the flag ironic_serial_console_autoenable,
which, when enabled, will configure serial consoles for each of the ironic
nodes during the `overcloud post configure` step. This was added to reduce the
number of kayobe commands that are needed to be run for a given deployment.

Change-Id: I7072e518c29387b964e4e25b08b4559f152d5ecf
Story: 2004192
Task: 29325