Work around Galaxy semver issues
Adding the tag 1.0.1 to the grafana-conf galaxy role caused 1.0 to no
longer be accessible. Changing the version format from a.b to a.b.c was
required to get Galaxy to pick up the new version. This change uses an
identical version of the role.

This commit needs to be included in stable/pike first for the
kayobe-overcloud-upgrade-centos job on stable/queens to succeed.

Change-Id: I51580b5df1496b95eaa3755b6d181511704ecbe8