Handle existing kolla-ansible inventory directories
Stable branches only

During the rocky release, kayobe is moving to using a directory for the
kolla-ansible inventories, rather than a file. The forward upgrade path
is handled, but downgrade is not. If using a single environment and
switching between releases, it's possible that a directory generated by
rocky will be present for queens, and cause an error.

This change deletes any inventory directories found.

Change-Id: Ia4e88a544d96eeb621135483693c2ff2ee9bf5d7
Story: 2004806
Task: 28961