Allow to configure where to deploy a Docker Registry
This commit changes the docker-registry playbook to deploy a Docker
Registry on hosts in the docker-registry group, rather than being
harcoded to deploy on the first controller.

The default location is changed to deploy on the seed as it appears to
be a good location for most deployments.

UpgradeImpact: The Docker Registry is now deployed on the seed by
default.

Change-Id: I87d02bd9556afb795c8783ed78e797fdd31fe8ca
Story: 2004820
Task: 28986