Rocky: Add missing monasca groups to inventory
Kolla is backporting all fixes for monasca to the stable/rocky branch.

This patch adds relevant groups to the inventory necessary to deploy
monasca. It should be compatible with versions kolla-ansible prior to
the monasca backport since the group changes are only additive.

We also disable grafana by default, instead use the monasca-grafana
service that is enabled automatically when monasca is enabled.

Change-Id: I35e2eb5bd96186537cfd085ecd611873bcd26801