Use BatchMode to check whether host is accessible via SSH
If password authentication is not disabled on the server, we may be
prompted for a password if the Kayobe Ansible user doesn't exist or
isn't configured for passwordless authentication.

By using BatchMode, the ssh command can fail quickly and proceed to
create the Kayobe Ansible user using the bootstrap user.

Change-Id: If22ed34dc4b6e87f8cf76c302948c955bddf2bc5