Provide keystone.conf to kolla-ansible by default
Commit d8ee0fa027140fb8a7504bd6685b6dcdfe3ac8fe introduced the variable
kolla_enable_keystone without setting a default value and used it to
check whether to template and provide keystone.conf to kolla-ansible,
unlike the extended keystone configuration which was always copied. This
resulted in non-intuitive behaviour if kolla_enable_keystone wasn't
explicitly set by the operator.

This commit sets kolla_enable_keystone to true by default, which matches
the default in kolla-ansible, and uses it instead of hardcoded true
values.

Change-Id: Id090348a6f22f251555fa286a7d24923087122a9
Story: 2005930
Task: 34273