Fix inspector iPXE boot with UEFI compute nodes
Rocky only

The IPA ramdisk fails to boot on UEFI compute nodes with iPXE. We see
the following error on the console, followed by a kernel panic:

Fix kernel panic-not syncing: VFS: unable to mount root fs on unknown
block(0,0)

The cause is the lack of an initrd=ironic-agent.initramfs kernel
argument in inspector.ipxe.

Rather than removing the kayobe override template as was done in stein,
for Rocky we take the less risky approach of adding the kernel argument
to our template.

Change-Id: I5363cb18b361d450ef8e77c553cc21ee2241ba79
Story: 2006214
Task: 35804